initial commit for this branch, partial progress
